+++ docs/release_notes.rst
@@ -2,6 +2,15 @@
 Release Notes
 =============
 
+0.9.4
+-----
+
+* Added ``JavetConverterConfig``
+* Added ``JavetEntityFunction`` for ``JavetObjectConverter``
+* Updated ``JavetObjectConverter`` to allow skipping functions
+* Added ``getPrototype()`` and ``setPrototype()`` to ``V8Runtime``
+* Changed the way that Javet libraries are deployed to be multi-process safe
+
 0.9.3
 -----